Kotlin API call
