+++ app/src/main/java/com/learning/androidlearning/movemarker/kotlin/splash/splashnew/UserDataAdapter.kt
@@ -1,11 +1,10 @@
-package com.learning.androidlearning.movemarker.kotlin.splash
+package com.learning.androidlearning.movemarker.kotlin.splash.splashnew
 
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
-import com.bumptech.glide.Glide
 import com.learning.androidlearning.R
 
 class UserDataAdapter(val userData: List<UserData>): RecyclerView.Adapter<UserDataViewHolder>(){
