+++ app/src/main/res/layout/activity_splash.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".movemarker.kotlin.splash.SplashActivity">
+    tools:context=".movemarker.kotlin.splash.splashnew.SplashActivity">
 
     <TextView
         android:id="@+id/tv_userdata"
